requirements.txt
psycopg2-binary
PyYAML
streamlit
pymongo
pandas
plotly
